requirements.txt
langchain
langchain-huggingface
faiss-cpu
streamlit
langchain-community
sentence-transformers
python-dotenv